converter/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "darkentrery",email = "badajkovmihail@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "alembic (>=1.14.1,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "reportlab (>=4.3.1,<5.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pdfkit (>=1.0.0,<2.0.0)",
    "alembic-postgresql-enum (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
